fix: resolve two Docker registry bugs preventing docker push
Bug 1: Cast upload_timeout config to int in both config/packgrid.php
and BlobStorageService to prevent Carbon TypeError on blob uploads.

Bug 2: Remove fallback in DockerRegistryAuth that returned unvalidated
short usernames (e.g. "token") as token values, causing 401 on blob
HEAD requests during docker push.

+++ app/Services/Docker/BlobStorageService.php
@@ -167,7 +167,7 @@ class BlobStorageService
 
     public function initChunkedUpload(DockerRepository $repository): DockerUpload
     {
-        $uploadTimeout = config('packgrid.docker.upload_timeout', 86400);
+        $uploadTimeout = (int) config('packgrid.docker.upload_timeout', 86400);
         $tempPath = $this->generateTempPath();
 
         // Ensure temp directory exists
